Programs/Python/dieScanner/diescanProd/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "diescanprod"
version = "1.0.0"
description = "Production die scanner package for FANUC robot forging cell"
authors = [{name = "USA Forge Cell Team", email = "team@usaforgecell.com"}]
license = {text = "Proprietary"}
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Manufacturing",
    "License :: Other/Proprietary License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Automation",
    "Topic :: System :: Hardware :: Robots",
]
dependencies = [
    "cpppo>=4.0.0",
    "numpy>=1.20.0",
    "pyserial>=3.5",
    "click>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=0.991",
]

[project.scripts]
diescan-prod = "diescanprod.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=src/diescanprod --cov-report=html --cov-report=term-missing"